+++ CHANGES.md
@@ -40,6 +40,8 @@ on going at master
 * Fix List
   - Fixed possible problem of poisson regression for R.
 * Python module now throw exception instead of crash terminal when a parameter error happens.
+* Python module now has importance plot and tree plot functions.
 * Java api is ready for use
 * Added more test cases and continuous integration to make each build more robust
 * Improvements in sklearn compatible module
+* Added pip installation functionality for python module